Update kaldi threading utilities to use STL thread support library instead of pthread.
