+++ src/util/kaldi-semaphore.h
@@ -1,6 +1,7 @@
-// thread/kaldi-semaphore.h
+// util/kaldi-semaphore.h
 
 // Copyright 2012  Karel Vesely (Brno University of Technology)
+//           2017  Dogan Can (University of Southern California)
 
 // See ../../COPYING for clarification regarding multiple authors
 //
@@ -21,36 +22,28 @@
 #ifndef KALDI_THREAD_KALDI_SEMAPHORE_H_
 #define KALDI_THREAD_KALDI_SEMAPHORE_H_ 1
 
-#include <pthread.h>
+#include <mutex>
+#include <condition_variable>
 
 namespace kaldi {
-  
+
 class Semaphore {
  public:
-  Semaphore(int32 initValue = 0); 
+  Semaphore(int32 count = 0);
+
   ~Semaphore();
 
-  bool TryWait(); ///< Returns true if Wait() goes through
-  void Wait(); ///< decrease the counter
-  void Signal(); ///< increase the counter
-  
-  /**
-   * returns the counter value, 
-   * zero means no resources, the Wait() will block
-   */ 
-  int32 GetValue() {
-    return counter_; 
-  }
+  bool TryWait();  ///< Returns true if Wait() goes through
+  void Wait();     ///< decrease the counter
+  void Signal();   ///< increase the counter
 
  private:
-  int32 counter_; ///< the semaphore counter, 0 means block on Wait() 
-  
-  pthread_mutex_t mutex_;
-  pthread_cond_t cond_;
+  int32 count_;    ///< the semaphore counter, 0 means block on Wait()
+
+  std::mutex mutex_;
+  std::condition_variable condition_variable_;
   KALDI_DISALLOW_COPY_AND_ASSIGN(Semaphore);
 };
-
-
 
 } //namespace
