+++ src/util/kaldi-table-inl.h
@@ -25,6 +25,7 @@
 
 #include <algorithm>
 #include <string>
+#include <thread>
 #include <utility>
 #include <vector>
 #include <errno.h>
@@ -32,8 +33,7 @@
 #include "util/kaldi-holder.h"
 #include "util/text-utils.h"
 #include "util/stl-utils.h"  // for StringHasher.
-#include "thread/kaldi-thread.h"
-#include "thread/kaldi-semaphore.h"
+#include "util/kaldi-semaphore.h"
 
 
 namespace kaldi {
@@ -715,18 +715,8 @@ class SequentialTableReaderBackgroundImpl:
     KALDI_ASSERT(base_reader_ != NULL &&
                  base_reader_->IsOpen());  // or code error.
     {
-      pthread_attr_t pthread_attr;
-      pthread_attr_init(&pthread_attr);
-      int32 ret = pthread_create(
-          &thread_,
-          &pthread_attr,
-          SequentialTableReaderBackgroundImpl<Holder>::run,
-          static_cast<void*>(this));
-      if (ret != 0) {
-        const char *c = strerror(ret);
-        KALDI_WARN << "Error creating thread, errno was: " << c;
-        return false;
-      }
+      thread_ = std::thread(SequentialTableReaderBackgroundImpl<Holder>::run,
+                            this);
     }
 
     if (!base_reader_->Done())
@@ -775,11 +765,8 @@ class SequentialTableReaderBackgroundImpl:
       return;
     }
   }
-  static void* run(void *object_in) {
-    SequentialTableReaderBackgroundImpl<Holder> *object =
-        reinterpret_cast<SequentialTableReaderBackgroundImpl<Holder>*>(object_in);
+  static void run(SequentialTableReaderBackgroundImpl<Holder> *object) {
     object->RunInBackground();
-    return NULL;
   }
   virtual bool Done() const {
     return key_.empty();
@@ -827,7 +814,7 @@ class SequentialTableReaderBackgroundImpl:
   // note: we can be sure that Close() won't be called twice, as the TableReader
   // object will delete this object after calling Close.
   virtual bool Close() {
-    KALDI_ASSERT(base_reader_ != NULL && KALDI_PTHREAD_PTR(thread_) != 0);
+    KALDI_ASSERT(base_reader_ != NULL && thread_.joinable());
     // wait until the producer thread is idle.
     consumer_sem_.Wait();
     bool ans = true;
@@ -842,10 +829,7 @@ class SequentialTableReaderBackgroundImpl:
     base_reader_ = NULL;
     producer_sem_.Signal();
 
-    if (pthread_join(thread_, NULL) != 0) {
-      KALDI_WARN << "Error rejoining thread.";
-      return false;
-    }
+    thread_.join();
     return ans;
   }
   ~SequentialTableReaderBackgroundImpl() {
@@ -864,7 +848,7 @@ class SequentialTableReaderBackgroundImpl:
   // that the producer (background thread) waits on.
   Semaphore consumer_sem_;
   Semaphore producer_sem_;
-  pthread_t thread_;
+  std::thread thread_;
   SequentialTableReaderImplBase<Holder> *base_reader_;
 
 };
